refactor(engine): extract order reconciliation from TradingEngine
strategy.py carried ~470 lines of order-reconciliation logic (startup
rehydration, recovering unknown bot orders, refreshing missing persisted orders,
and drift escalation) as engine methods. Move them into
engines/order_reconciliation.py as engine-taking functions (3401->2951),
following the established collaborator-function pattern; the engine keeps thin
delegating methods so all call sites (init, main loop, tests) are unchanged.

Behavior-preserving: the state-recovery and reconciliation/drift tests, the full
1225-test live_trade suite, and mypy all pass.
